migrate `prove_clone` to a "base impls" test file
