+++ src/test.rs
@@ -112,43 +112,7 @@ fn solve_goal(program_text: &str, goals: Vec<(&str, SolverChoice, &str)>) {
     }
 }
 
-#[test]
-fn prove_clone() {
-    test! {
-        program {
-            struct Foo { }
-            struct Bar { }
-            struct Vec<T> { }
-            trait Clone { }
-            impl<T> Clone for Vec<T> where T: Clone { }
-            impl Clone for Foo { }
-        }
-
-        goal {
-            Vec<Foo>: Clone
-        } yields {
-            "Unique; substitution [], lifetime constraints []"
-        }
-
-        goal {
-            Foo: Clone
-        } yields {
-            "Unique; substitution [], lifetime constraints []"
-        }
-
-        goal {
-            Bar: Clone
-        } yields {
-            "No possible solution"
-        }
-
-        goal {
-            Vec<Bar>: Clone
-        } yields {
-            "No possible solution"
-        }
-    }
-}
+mod basic_impls;
 
 #[test]
 fn inner_cycle() {
