+++ src/test/basic_impls.rs
@@ -1,0 +1,39 @@
+use super::*;
+
+#[test]
+fn prove_clone() {
+    test! {
+        program {
+            struct Foo { }
+            struct Bar { }
+            struct Vec<T> { }
+            trait Clone { }
+            impl<T> Clone for Vec<T> where T: Clone { }
+            impl Clone for Foo { }
+        }
+
+        goal {
+            Vec<Foo>: Clone
+        } yields {
+            "Unique; substitution [], lifetime constraints []"
+        }
+
+        goal {
+            Foo: Clone
+        } yields {
+            "Unique; substitution [], lifetime constraints []"
+        }
+
+        goal {
+            Bar: Clone
+        } yields {
+            "No possible solution"
+        }
+
+        goal {
+            Vec<Bar>: Clone
+        } yields {
+            "No possible solution"
+        }
+    }
+}
